Section position

+++ Form Builder/Models/Entity Property Files/Widgets/Checkbox/CheckboxSectionWidget+CoreDataProperties.swift
@@ -24,7 +24,7 @@ extension CheckboxSectionWidget {
     }
     
     /// CheckboxSectionWidget  convenience init
-    convenience init(title: String?, position: Int16) {
+    convenience init(title: String?, position: Int) {
         self.init(title: title, position: position, type: WidgetType.checkboxSectionWidget.rawValue)
     }
 }
