+++ Form Builder/Models/Entity Property Files/Widgets/Checkbox/CheckboxWidget+CoreDataProperties.swift
@@ -20,7 +20,7 @@ extension CheckboxWidget {
     @NSManaged public var checkboxSection: CheckboxSectionWidget?
 
     /// CheckboxWidget  convenience init
-    convenience init(title: String?, position: Int16) {
+    convenience init(title: String?, position: Int) {
         self.init(title: title, position: position, type: WidgetType.checkboxWidget.rawValue)
         self.checked = false
     }
