+++ Form Builder/Models/Entity Property Files/Widgets/Dropdown/DropdownWidget+CoreDataProperties.swift
@@ -19,7 +19,7 @@ extension DropdownWidget {
     @NSManaged public var dropdownSection: DropdownSectionWidget?
 
     /// DropdownWidget  convenience init
-    convenience init(title: String?, position: Int16) {
+    convenience init(title: String?, position: Int) {
         self.init(title: title, position: position, type: "DropdownWidget")
     }
 }
