+++ Form Builder/Models/Entity Property Files/Widgets/NumberFieldWidget+CoreDataProperties.swift
@@ -19,7 +19,7 @@ extension NumberFieldWidget {
     @NSManaged public var number: String?
     
     /// NumberFieldWidget  convenience init
-    convenience init(title: String?, position: Int16, number: String?) {
+    convenience init(title: String?, position: Int, number: String?) {
         self.init(title: title, position: position, type: WidgetType.numberFieldWidget.rawValue)
         self.number = number
     }
