+++ Form Builder/Models/Entity Property Files/Widgets/TextEditorWidget+CoreDataProperties.swift
@@ -19,7 +19,7 @@ extension TextEditorWidget {
     @NSManaged public var text: String?
     
     /// TextEditorWidget  convenience init
-    convenience init(title: String?, position: Int16, text: String?) {
+    convenience init(title: String?, position: Int, text: String?) {
         self.init(title: title, position: position, type: WidgetType.textEditorWidget.rawValue)
         self.text = text
     }
